Re-arranging files
